requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
databases==0.6.2
asyncpg==0.27.0
mido==1.2.10
sqlalchemy==1.4.47
python-multipart==0.0.6
tensorflow
opencv-python-headless==4.5.3.56
numpy==1.21.2
tensorflow-hub==0.12.0
absl-py==0.13.0
pydub
midi2audio
cv2

